chore: organize stories according to new linting rule
