+++ apps/storybook/src/ui/components/Avatar.stories.tsx
