+++ apps/storybook/src/ui/components/Paper.stories.tsx
